Server.New now starts listener goroutine
Remember how a couple of commits back I stopped allowing the caller to
close the client listener? Well that was a terrible idea, so instead
we're just gonna kick off the server in one go, and provide a Close()
method for cleaning up the server. This is good because we can ensure
that the server stops accepting connections, send a warning/QUIT
message to all the connected clients, and we can do some potential
logging in the future.