Refactor test of client registration
We don't actually have to check the state of the server after every
action, we can just send all the requests we need from the client and
then check the server's state at the end to make sure everything is
correct. This also removes the need for the weird channel/goroutine
polling system (which was racy).