Use contexts for handling Server/Client shutdown
Each client creates a new context from the server's context, so that if
the server has to shutdown for some reason, the canceling of the
server's context will trigger the cancelation of all the clients.
Contexts are cool.

In this commit, we also send QUIT to all of a client's connected
channels if they leave without sending their own quit mesage (either
they were kicked off, or they disconnected suddenly).