Queue messages as they arrive from clients
The previous design was racy because messages would always get executed
immediately for each client with no synchronization between client
goroutines. The solution I chose is to have each client goroutine send
a message to another goroutine that listens on the channel and executes
messages in sequence. Though this channel is buffered, one potential
flaw with this design is that if any one particular messages takes a
non-trivial amount of time to complete, then it will be stuck blocking
and in turn block all the connected clients from sending more messages.

We also needed to add synchronization for when clients arrive/leave.
These also needed to be synchronized to message execution, since there
are messages which mofidy the client/channel lists.

Because all synchronization is now relegated to the server, we can
remove the locks from util.List.

One last thing: we have created a goroutine for each client goroutine
that is responsible for trying to read the next input from the client.
This is because the client may be blocked on a read that is taking some
amount of time, and drop connection halfway through; we need to be able
to select between the client canceling and a new input.