Use maps instead of util.List
maps are more scalable because of faster lookup times, which I realized
could become a real problem with the array based approach we had
previously. As a consequence of using the NICK as keys, clients are
now only officially added to the server once they have completed the
registration process, which is probably how it should've been all
along. Once PING/PONG or some other timeout mechanism is implemented,
straggler clients who have not been registered will be able to be
cleaned up.