Replace queuing msgPair with a plain function
Instead of creating a struct with the requisite information, just send
the msgQueue a func that describes what to do. We also now queue up
client cancellation, which allows us to remove the msgLock from the
Server entirely. Everything now executes sequentially, which may end up
being a performance issue in the future, but I didn't like the current
solution because it mixed locks with the channel queuing. I think a
RWLock solution may be the way to go moving forward.