Parse mode strings in sequential order
Instead of returning two sets of addition/subtraction slices of
modeChars to apply, return a struct that contains all that information.
This is necessary because channel modes have parameters, and it is
important to keep the modeString in the correct order as the parameters
are given in the same order as they appear in in the modeStr.