Don't treat nospcrlfcl as a terminal
Let the parser handle this, not the lexer. This is needed because
message tags have different rules about the kind of characters that can
be contained in them, so lexing everything as the same type of lexeme
doesn't make much sense. This also has the added benefit of "cleaning"
up source parsing, since we can now delineate between special
characters like '!' and '@'.