Stop excess goroutines after Close
The connection and msgQueue goroutines were leaking after Close, now
they should stop along with the Server. We also delegate client
creation and handling to the connection goroutine to free up tasks from
the main thread, which now only concerns itself with stopping when it
receives an interrupt or is cancel()ed.