Use a strings.Builder for nospcrlfcl
This was a particularly memory heavy function upon profiling, and
having this new buffer greatly improves memory usage. It may be
worthwhile revisiting msg/parse.go and doing the same for the rest of
the string concatenations, but those are much lower on the memory
usage (and on the priority list).