Add flood control
Each client gets a pool of 10 grants. Every 2 seconds, they are given
another grant. In order to read a message from a client, the server
must request a grant from the client. If the server tries to read from
the client and they have no grants left, they are kicked off for
flooding.

Because of this change, Client.New now accepts a context in order to
stop the granting goroutine. Some of the exec tests also had to be
changed to give more grants.