Remove server message queue pattern
By introducing locks for the server's client and channel maps, each
client can now coordinate with each other without the need for the
server to make sure messages execute in the correct order.

Another benefit of this is for AUTHENTICATE; previously, the server
would slow down because the main goroutine was in charge of running
expensive cryptographic functions like bcrypt. Now that is offloaded
to each of the client goroutines.